[Fix] #8 - 코드 컨벤션 준수에 따른 코드수정
