+++ SKHUAZ/SKHUAZ/Util/Extension/UIStackView++.swift
@@ -13,4 +13,13 @@ extension UIStackView {
             self.addArrangedSubview(view)
         }
     }
+    
+    func removeAllArrangedSubviews() {
+        let removedSubviews = arrangedSubviews.reduce([]) { (allSubviews, subview) -> [UIView] in
+            self.removeArrangedSubview(subview)
+            return allSubviews + [subview]
+        }
+        NSLayoutConstraint.deactivate(removedSubviews.flatMap({ $0.constraints }))
+        removedSubviews.forEach({ $0.removeFromSuperview() })
+    }
 }
